Not bad...
Should now handle VersionTagPlugin MaxMajor
configuration and StartingVersion or may be better
a MinMajor to avoid users to use StartingVersion
instead of +Fake versions.

Before that, relationships between +Fake and hiding tags
must be clarified. Today, a hiding tag hides a Fake. This seems
fine: this allow a Fake to be cleaned up. But do we want this?
This is a highly manual operation, so yes. We want this.
It seems that +Fake should just be correctly handled by Rebuild:
they must be ignored.